Remove the .ci directory.
As part of this transition, we are going to be making a lot of changes
to CI. Once all of those changes are made, we might look into automating
some of the CI yaml scripts. For now, these changes will be manual, and
the file used for yaml generation isn't needed.